+++ src/@next/EmptyState/assets/images/emptyCarton.ts
