+++ src/@next/EmptyState/assets/images/emptyMailbox.ts
