+++ src/@next/EmptyState/assets/images/index.ts
@@ -1,0 +1,3 @@
+export { default as emptyCarton } from './emptyCarton';
+export { default as emptyMailbox } from './emptyMailbox';
+export { default as safetyCone } from './safetyCone';
