+++ src/@next/EmptyState/assets/images/safetyCone.ts
